vmalert: avoid sending resolves for pending alerts
Before the change we were sending notifications to notifier
if:
* alert is in Fire state
* alert is in Inactive state
We were sending Inactive notifications with End set to
last execution time to resolve alert ASAP. Unfortunately,
we were sending resolves for Pending alerts that become
Inactive, which is wrong.
In this change we delete alert from the active list if
it was Pending and become Inactive. In this way we now
have Inactive alerts only if they were in state Fire before.
See test change for example.