vmselect: add query tracing feature
Query tracing feature allows displaying the stages of query processing.
The stages can be then visualized in order to understand the slowest
or the most expensive parts of the query processing.

The tracing is implemented via `lib/tracer`. The library allows creating
a tracing context, which supports nesting. The context is then used
in `vmselect` and `lib/storage` in order to highlight the most interesting
parts of query processing.

The query tracing can be enabled for each specific query by passing extra
GET params in request. The trace will be then collected and printed to
the stdout and also attached to the response (only `/api/v1/query` and
`api/v1/query_range` support tracing in response).

See more at `tracer/example_context_test.go`

Signed-off-by: hagen1778 <roman@victoriametrics.com>