logger: redact password for URL flags
It is possible to leak passwords and similar secrets for URL flags, e.g.
`remoteWrite.url`, when a user:pass is supplied, as this is logged out
by the flag logger.

To reduce the chance of misuse, add some handling for URL flags that
redacts the password prior to logging, if present.