feat: enable gzip for streamParse mode
Up until now, targets parsed in streamParse mode were not scraped with
the `Accept-Encoding gzip` header. This was particularly noticeable when
the target's response size exceeded maxScrapeSize.
With this commit, the behaviour is now the same for streamParse mode as
it is for "normal" mode: the http body is read, its size compared to
maxScrapeSize, and then the data is ungzipped before being processed.