lib/logger: add failing test-case for `LogThrottler` If the goroutine is blocked on `<-lt.ch`, it means there have been no error/warn calls in this throttle period yet.
When the first error/warn call is made, the channel is empty, so it
successfully sends to the channel and then logs the message. The
goroutine's receive then unblocks and empties the channel before
going to sleep, which allows a *second* error/warn call to succeed
immediately, rather than being discarded.

Note that there is a race if the two calls are made immediately after
creating the `LogThrottler`, and the goroutine has not yet reached its
first channel-receive. In this case, the second call will be discarded
correctly, masking the bug.