vmalert: fix nil map assignment
The storage instance with nil map params was created for remote-read purposes.
And before change https://github.com/VictoriaMetrics/VictoriaMetrics/pull/4341/commits/7a9ae9de0d2221cf7759e49ebcde0a06508b0f68 this map was ignored in ApplyParams.
Now, it started to be used and vmalert panics in runtime.

The fix properly inits map for at `NewVMStorage` and verifies it is not nil
on assignment in `ApplyParams`.

Signed-off-by: hagen1778 <roman@victoriametrics.com>