lib/mergeset: user RWMutex for parts access
The `partsLock` mutex can be acquired by merging workers
and during the index search. On the systems with high number
of CPUs or high `search.maxConcurrentRequests` value this could
result in contention on access to parts, preventing either reads
or merges.

This change is an attempt to reduce the access contention to parts.

Signed-off-by: hagen1778 <roman@victoriametrics.com>