app/vmalert: follow-up after c9375cac5e7bc612259e0101b50b4861115cc030
https://github.com/VictoriaMetrics/VictoriaMetrics/commit/c9375cac5e7bc612259e0101b50b4861115cc030

Descriptions were updated in attempt to make it more clear for readers,
re-phrasing and linking missing docs.

`eval_delay` was added to tests to verify it can be unmarshalled.

`eval_delay` is now applied before timestamp alignment to make it more predictable.
Before, if delay < interval the timestamp won't be aligned.

`eval_delay` and `eval_offset` was added to API output.

`PreviouslySentSeriesToRW` converted to private `previouslySentSeriesToRW`.

Signed-off-by: hagen1778 <roman@victoriametrics.com>