add custom compressor for time series for deduplication
Currently, for deduplicating vmagents needs to keep
list of time series keys in memory.
The key is marshalled into string labels. On avg, such key
will occupy about 100B in mem.

In order to reduce the size of the key, this PR adds a custom
compressor which would break a time series in label-value pairs
and produce the hash (4B) of the pair. Then key could be just
a concat of the hashes. This should reduce the key size.

Signed-off-by: hagen1778 <roman@victoriametrics.com>