remotewrite: Use TLS ServerName as Host in all requests
This change modifies HTTP requests made via remote-write to set the
`Host` field to the TLS config's `ServerName`, when a TLS config is
present. Otherwise, the empty default is maintained, and the Go stdlib
will fall back to using the host component of the URL.

This is useful to work around a somewhat-unfortunate setup we are
currently dealing with, involving:

* a client machine that has limited ability to resolve DNS queries (so
  we must point to our reverse-proxy instance via IP)
* a reverse proxy that proxies based on the request's target host

We can almost get this to work by passing both `-remoteWrite.url` and
`-remoteWrite.tlsServerName`; however, the host will still be the host
specified in the former and not the latter, which our reverse proxy
mappings will not be able to handle.

---

This PR as-is is likely not production-worthy:

* this may not be the best place/way to get the TLS config and set the
  Host field
* more similar instances may need to be modified across the codebase for
  consistency

With this PR, I'm looking to get feedback on:

* which behavior is the most correct/useful (Host field comes from URL
  vs Host field comes from TLS ServerName)
* if this change is desired, then particulars on how would be
  appreciated!